Reformatted and cleaned project

+++ tests/UltraExtremeTest/src/ultraextreme/model/item/BasicWeaponTest.java
@@ -36,15 +36,18 @@ public class BasicWeaponTest extends TestCase {
 	public void testFireShot() {
 		float epsilon = 0.001f;
 		assertTrue(bulletManager.getBullets().size() == 0);
-		float cooldown = (float)BasicWeapon.getInitCooldown();
+		float cooldown = BasicWeapon.getInitCooldown();
 
-		basicWeapon.fire(new Position(), PlayerID.PLAYER1, new Rotation(0), cooldown * (1 + epsilon));
+		basicWeapon.fire(new Position(), PlayerID.PLAYER1, new Rotation(0),
+				cooldown * (1 + epsilon));
 		assertTrue(bulletManager.getBullets().size() == 1);
 
-		basicWeapon.fire(new Position(), PlayerID.PLAYER1, new Rotation(0), cooldown * (1 + epsilon));
+		basicWeapon.fire(new Position(), PlayerID.PLAYER1, new Rotation(0),
+				cooldown * (1 + epsilon));
 		assertTrue(bulletManager.getBullets().size() == 2);
 
-		basicWeapon.fire(new Position(), PlayerID.PLAYER1, new Rotation(0), cooldown * (1 + epsilon));
+		basicWeapon.fire(new Position(), PlayerID.PLAYER1, new Rotation(0),
+				cooldown * (1 + epsilon));
 		assertTrue(bulletManager.getBullets().size() == 3);
 	}
 
@@ -52,9 +55,9 @@ public class BasicWeaponTest extends TestCase {
 	 * Test so the properties of the bullet that the weapon fires are correct.
 	 */
 	public void testBulletProperties() {
-		float cooldown = (float)BasicWeapon.getInitCooldown();
+		float cooldown = BasicWeapon.getInitCooldown();
 		basicWeapon.fire(new Position(10, 5), PlayerID.PLAYER1,
-				new Rotation(0), cooldown + cooldown/1000);
+				new Rotation(0), cooldown + cooldown / 1000);
 		IBullet bullet = bulletManager.getBullets().get(0);
 		assertTrue(bullet instanceof BasicBullet);
 		assertEquals(bullet.getWidth(), 5);
